Move screens to their own package
