+++ app/src/androidTest/java/com/hedvig/app/feature/offer/ConcurrentInceptionTest.kt
@@ -2,6 +2,8 @@ package com.hedvig.app.feature.offer
 
 import com.hedvig.android.owldroid.graphql.OfferQuery
 import com.hedvig.app.feature.embark.masking.ISO_8601_DATE
+import com.hedvig.app.feature.offer.screen.ChangeDateSheet
+import com.hedvig.app.feature.offer.screen.OfferScreen
 import com.hedvig.app.feature.offer.ui.OfferActivity
 import com.hedvig.app.testdata.feature.offer.BUNDLE_WITH_CONCURRENT_INCEPTION_DATES_SPECIFIC_DATE
 import com.hedvig.app.testdata.feature.offer.CONCURRENT_INCEPTION_START_DATE
