+++ app/src/androidTest/java/com/hedvig/app/feature/offer/MultipleNonSwitchableSwitcherTest.kt
@@ -2,6 +2,7 @@ package com.hedvig.app.feature.offer
 
 import com.hedvig.android.owldroid.graphql.OfferQuery
 import com.hedvig.app.R
+import com.hedvig.app.feature.offer.screen.OfferScreen
 import com.hedvig.app.feature.offer.ui.OfferActivity
 import com.hedvig.app.testdata.feature.offer.OFFER_DATA_NORWAY_BUNDLE_HOME_CONTENTS_TRAVEL_MULTIPLE_PREVIOUS_INSURERS
 import com.hedvig.app.util.ApolloCacheClearRule
