+++ app/src/androidTest/java/com/hedvig/app/feature/offer/MultipleQuotesTest.kt
@@ -1,6 +1,9 @@
 package com.hedvig.app.feature.offer
 
 import com.hedvig.android.owldroid.graphql.OfferQuery
+import com.hedvig.app.feature.offer.screen.FAQBottomSheetScreen
+import com.hedvig.app.feature.offer.screen.OfferScreen
+import com.hedvig.app.feature.offer.screen.QuoteDetailScreen
 import com.hedvig.app.feature.offer.ui.OfferActivity
 import com.hedvig.app.feature.perils.PerilRecyclerItem
 import com.hedvig.app.testdata.feature.offer.OFFER_DATA_NORWAY_BUNDLE_HOME_CONTENTS_TRAVEL
