+++ app/src/androidTest/java/com/hedvig/app/feature/offer/StartDateFromPreviousInsurerTest.kt
@@ -1,8 +1,9 @@
 package com.hedvig.app.feature.offer
 
-import com.agoda.kakao.screen.Screen.Companion.onScreen
 import com.hedvig.android.owldroid.graphql.OfferQuery
 import com.hedvig.app.R
+import com.hedvig.app.feature.offer.screen.ChangeDateView
+import com.hedvig.app.feature.offer.screen.OfferScreen
 import com.hedvig.app.feature.offer.ui.OfferActivity
 import com.hedvig.app.testdata.feature.offer.BUNDLE_WITH_START_DATE_FROM_PREVIOUS_INSURER
 import com.hedvig.app.util.ApolloCacheClearRule
