+++ app/src/androidTest/java/com/hedvig/app/feature/offer/TodayStartDateTest.kt
@@ -2,6 +2,7 @@ package com.hedvig.app.feature.offer
 
 import com.hedvig.android.owldroid.graphql.OfferQuery
 import com.hedvig.app.R
+import com.hedvig.app.feature.offer.screen.OfferScreen
 import com.hedvig.app.feature.offer.ui.OfferActivity
 import com.hedvig.app.testdata.feature.offer.BUNDLE_WITH_CONCURRENT_INCEPTION_DATES
 import com.hedvig.app.util.ApolloCacheClearRule
