+++ app/src/androidTest/java/com/hedvig/app/feature/offer/screen/OfferScreen.kt
@@ -1,15 +1,12 @@
-package com.hedvig.app.feature.offer
+package com.hedvig.app.feature.offer.screen
 
 import android.view.View
-import com.agoda.kakao.check.KCheckBox
 import com.agoda.kakao.common.views.KView
 import com.agoda.kakao.recycler.KRecyclerItem
 import com.agoda.kakao.recycler.KRecyclerView
-import com.agoda.kakao.text.KButton
 import com.agoda.kakao.text.KTextView
 import com.hedvig.app.R
 import com.hedvig.app.feature.offer.ui.OfferActivity
-import com.hedvig.app.feature.offer.ui.changestartdate.ChangeDateBottomSheet
 import com.hedvig.app.util.withParentIndex
 import com.kaspersky.kaspresso.screens.KScreen
 import org.hamcrest.Matcher
@@ -21,13 +18,13 @@ object OfferScreen : KScreen<OfferScreen>() {
     val scroll = KRecyclerView(
         { withId(R.id.offerScroll) },
         {
-            itemType(::HeaderItem)
-            itemType(::SwitcherItem)
-            itemType(::Facts)
-            itemType(::QuoteDetail)
-            itemType(::FAQ)
-            itemType(::InfoCard)
-            itemType(::WarningCard)
+            itemType(OfferScreen::HeaderItem)
+            itemType(OfferScreen::SwitcherItem)
+            itemType(OfferScreen::Facts)
+            itemType(OfferScreen::QuoteDetail)
+            itemType(OfferScreen::FAQ)
+            itemType(OfferScreen::InfoCard)
+            itemType(OfferScreen::WarningCard)
         }
     )
 
@@ -68,21 +65,3 @@ object OfferScreen : KScreen<OfferScreen>() {
     }
 }
 
-object ChangeDateSheet : KScreen<ChangeDateSheet>() {
-    override val layoutId = R.layout.dialog_change_start_date
-    override val viewClass = ChangeDateBottomSheet::class.java
-    val submit = KButton { withId(R.id.chooseDateButton) }
-    val changeDateContainer = KView {
-        withId(R.id.change_date_container)
-    }
-}
-
-object ChangeDateView : KScreen<ChangeDateView>() {
-    override val layoutId = R.layout.change_date
-    override val viewClass = ChangeDateView::class.java
-
-    val switches = KCheckBox {
-        withId(R.id.auto_set_date_switch)
-        isDisplayed()
-    }
-}
